fix(src): Sistemati metodi Local Similrity
Sistemati metodi Local Sim per rispettare la nuova struttura della
libreria.

Aggiornata documentazione
