--- cnlp/similarity_methods/local_similarity/hub_depressed.py
@@ -1,52 +1,0 @@
-import networkx as nx
-import numpy as np
-from .common_neighbors import __common_neighbors
-from cnlp.utils import nodes_to_indexes
-from scipy.sparse import lil_matrix, csr_matrix
-
-
-def __hub_depressed(G: nx.Graph, x, y) -> float:
-    """Compute the Hub Depressed Index for 2 given nodes."""
-    return __common_neighbors(G, x, y) / max(G.degree[x], G.degree[y])
-
-
-def hub_depressed(G: nx.Graph) -> csr_matrix:
-    """Compute the Hub Depressed Index for all nodes in the Graph.
-    Each similarity value is defined as:
-
-    .. math::
-        S(x, y) = \\frac{2 |\\Gamma(x) \\cap \\Gamma(y)|}{\\max(k_x, k_y)}
-
-    where \\(\\Gamma(x)\\) are the neighbors of node \\(x\\)
-    and \\(k_x\\) is the degree of the node \\(x\\).
-
-
-    Parameters
-    ----------
-    G: nx.Graph :
-        input Graph (a networkx Graph)
-
-    Returns
-    -------
-    S: csr_matrix : the Similarity Matrix (in sparse format)
-
-    Notes
-    -----
-    This index is the same as the previous one but with the
-    opposite goal as it avoids the formation of
-    links between hubs and low degree nodes in the networks.
-    The Hub depressed index promotes the links evolution
-    between the hubs as well as the low degree nodes.
-    """
-    size = G.number_of_nodes()
-    S = lil_matrix((size, size))
-    node_index_map = nodes_to_indexes(G)
-
-    for x, y in nx.complement(G).edges():
-        _x = node_index_map[x]
-        _y = node_index_map[y]
-
-        S[_x, _y] = __hub_depressed(G, x, y)
-        # S[y, x] = S[x, y]
-
-    return S.tocsr()
