+++ docs/similarity_methods/index.html
@@ -65,7 +65,7 @@ from . import global_similarity</code></pre>
 <dd>
 <div class="desc"><p>Collection of Global Similarity Methods for Link Prediction …</p></div>
 </dd>
-<dt><code class="name"><a title="cnlp.similarity_methods.local_similarity" href="local_similarity/index.html">cnlp.similarity_methods.local_similarity</a></code></dt>
+<dt><code class="name"><a title="cnlp.similarity_methods.local_similarity" href="local_similarity.html">cnlp.similarity_methods.local_similarity</a></code></dt>
 <dd>
 <div class="desc"><p>Collection of Local Similarity Methods for Link Prediction …</p></div>
 </dd>
@@ -96,7 +96,7 @@ from . import global_similarity</code></pre>
 <li><h3><a href="#header-submodules">Sub-modules</a></h3>
 <ul>
 <li><code><a title="cnlp.similarity_methods.global_similarity" href="global_similarity/index.html">cnlp.similarity_methods.global_similarity</a></code></li>
-<li><code><a title="cnlp.similarity_methods.local_similarity" href="local_similarity/index.html">cnlp.similarity_methods.local_similarity</a></code></li>
+<li><code><a title="cnlp.similarity_methods.local_similarity" href="local_similarity.html">cnlp.similarity_methods.local_similarity</a></code></li>
 <li><code><a title="cnlp.similarity_methods.quasi_local_similarity" href="quasi_local_similarity/index.html">cnlp.similarity_methods.quasi_local_similarity</a></code></li>
 </ul>
 </li>
